actually only keep lubs
This can only cause a problem in dimensions >= 3, where it is possible to find
an upperbound which is not a least upperbound by "climbing up" beside the bounds.

Imagine a structure like

X X U U U X X
 X X U U X X
  X X S X X

where we start from the node S and move our way upwards through ordinary nodes X
and upper bounds U. We will also find the nodes in the top line, which we do not
want to find.